third_party/blink/web_tests/svg/custom/createelement-expected.txt

CONSOLE ERROR: Uncaught TypeError: Cannot read properties of null (reading 'setAttribute')
